fix(admin): 통계 탭 500 — gold/exp/fame int 오버플로(::int → ::bigint) (#1019)
증상: 관리자 "통계" 탭이 전혀 로드되지 않음(HTTP 500 → "로드 실패").

원인: /api/admin/stats 의 raw SQL 이 character.v2 의 gold/exp/fame 를 `::int` 로
캐스팅하는데, gold 는 누적값이라 int4 범위(2,147,483,647)를 넘는 유저가 존재
(라이브 DB 실측: 한 유저 gold=111,749,639,926 ≈ 1117억). Postgres 가
`value "111749639926" is out of range for type integer` 로 쿼리 전체를 실패시켜
500 → 탭 전체가 깨짐.

수정: gold/exp/fame + kills/battleLosses 캐스팅을 `::bigint` 로. level 은 100 캡이라
int 유지. 응답 매핑(Number())·AdminStatsRow 타입 불변(bigint 는 node-pg 가 문자열로
반환 → Number() 가 안전 변환, 1117억 « 2^53). 라이브 DB 에 패치 쿼리 실측 — 정상
반환 확인. tsc 0 · eslint 0.

+++ src/app/api/admin/stats/route.ts
@@ -31,16 +31,18 @@ export async function GET() {
       p.name AS name,
       p.last_seen_at AS last_seen_at,
       (c.value->>'level')::int AS level,
-      (c.value->>'exp')::int AS exp,
-      (c.value->>'gold')::int AS gold,
-      (c.value->>'fame')::int AS fame,
+      -- exp/gold/fame 는 누적값이라 int4 범위(~21억)를 넘을 수 있다(예: gold 1000억+).
+      -- int 캐스팅은 "integer out of range" 로 쿼리 전체를 500 시키므로 bigint 로 받는다.
+      (c.value->>'exp')::bigint AS exp,
+      (c.value->>'gold')::bigint AS gold,
+      (c.value->>'fame')::bigint AS fame,
       (
         COALESCE((
-          SELECT SUM((m.value->>'kills')::int)
+          SELECT SUM((m.value->>'kills')::bigint)
           FROM jsonb_each(l.value->'monsters') AS m
           WHERE (m.value->>'kills') IS NOT NULL
         ), 0)
-        + COALESCE((l.value->>'battleLosses')::int, 0)
+        + COALESCE((l.value->>'battleLosses')::bigint, 0)
       ) AS battle_count
     FROM users u
     LEFT JOIN presence p ON p.user_id = u.id
